Update README.md
I was able to have the players be able to put their own names in the game. I was able able to have the board be 4x4, which looks really good when I ran it. I made it simple just so I was able to get the basics of it so that way I can just focus on having the code run and works well.